Links mongoDB with the index page
