+++ package.json
@@ -20,8 +20,11 @@
   },
   "homepage": "https://github.com/pushdev-code/moramia-website#readme",
   "dependencies": {
+    "body-parser": "^1.19.0",
     "ejs": "^3.1.2",
-    "express": "^4.17.1"
+    "envdot": "0.0.3",
+    "express": "^4.17.1",
+    "mongoose": "^5.9.19"
   },
   "devDependencies": {
     "@babel/core": "^7.9.6",
