+++ src/views/partials/content.ejs
@@ -1,38 +1,13 @@
 <main class="products-wrapper">
-    <article class="product">
-        <figure class="figure">
-            <img src="./img/product1.jpg" alt="Frutos secos">
-            <figcaption class="product-content">
-                <p class="product-price">From $180</p>
-                <h4 class="product-title">Frutos secos</h4>
-            </figcaption>
-        </figure>
-    </article>
-    <article class="product">
-        <figure class="figure">
-            <img src="./img/product2.jpg" alt="Nueces">
-            <figcaption class="product-content">
-                <p class="product-price">From $180</p>
-                <h4 class="product-title">Nueces</h4>
-            </figcaption>
-        </figure>
-    </article>
-    <article class="product">
-        <figure class="figure">
-            <img src="./img/product3.jpg" alt="Frutos secos">
-            <figcaption class="product-content">
-                <p class="product-price">From $180</p>
-                <h4 class="product-title">Frutos secos</h4>
-            </figcaption>
-        </figure>
-    </article>
-    <article class="product">
-        <figure class="figure">
-            <img src="./img/product4.jpg" alt="Frutos secos">
-            <figcaption class="product-content">
-                <p class="product-price">From $180</p>
-                <h4 class="product-title">Frutos secos</h4>
-            </figcaption>
-        </figure>
-    </article>
+  <%products.forEach( function( prd ){ %>
+  <article class="product">
+    <figure class="figure">
+      <img src=<%= prd.product_image %> alt=<%= prd.product_name %> >
+      <figcaption class="product-content">
+        <p class="product-price">From $<%= prd.product_price%></p>
+        <h4 class="product-title"><%= prd.product_name %></h4>
+      </figcaption>
+    </figure>
+  </article>
+  <% }) %>
 </main>
